corrections

+++ BS.Domain/Visitante.cs
@@ -9,7 +9,7 @@ namespace BS.Domain
         public int Cedula { get; set; }
         public string Nombre { get; set; }
         public string Email { get; set; }
-        public int Telefono { get; set; }
+        public long Telefono { get; set; }
         public bool Procesado { get; set; }
     }
 }
